Added instructions for rapid development iteration
This allows for building and running the binary directly in the
container to save rebuilding & redeploying the container image while
iterating quickly.

* It works on my machine
* Mileage may vary

This change works much better if all of the build scripts include `-j10`
or some reasonable value for CPUs on the calls to `make`, but I have
left them out of these patches for brevity.